+++ examples/tcp/server.rs
@@ -1,7 +1,7 @@
 use super::common::{Message};
 
 use message_io::events::{EventQueue};
-use message_io::network::{NetworkManager, NetEvent, Endpoint};
+use message_io::network::{Network, NetEvent, Endpoint};
 
 use std::collections::{HashMap};
 
@@ -17,8 +17,7 @@ pub fn run() {
     let mut event_queue = EventQueue::new();
 
     let network_sender = event_queue.sender().clone();
-    let mut network =
-        NetworkManager::new(move |net_event| network_sender.send(Event::Network(net_event)));
+    let mut network = Network::new(move |net_event| network_sender.send(Event::Network(net_event)));
 
     let mut clients: HashMap<Endpoint, ClientInfo> = HashMap::new();